refactor(examples): 迁移 snapline 至 plugins 目录
